pyproject.toml
[tool.poetry]
name = "foliant"
version = "1.0.12"
description = "Modular, Markdown-based documentation generator that makes pdf, docx, html, and more."
license = "MIT"
authors = ["Konstantin Molchanov <moigagoo@live.com>"]
readme = "README.md"
homepage = "https://foliant-docs.github.io/docs/"
repository = "https://github.com/foliant-docs/foliant"
documentation = "https://foliant-docs.github.io/docs/"
keywords = ["documentation", "markdown", "html", "docx", "pdf"]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.1.1"
cliar = "^1.3.2"
prompt_toolkit = "^2.0"

[tool.poetry.dev-dependencies]
pytest = "^3.6"
pytest-datadir = "^1.0"
pylint = "^2.5"
pytest-cov = "^2.5"
codecov = "^2.0"

[tool.poetry.scripts]
foliant = 'foliant.cli:entry_point'
